+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 - [Fixed issue where drag'n'dropped files were not uploaded in cloud
   projects.][11014]
 - [Fixed files associations not properly registered on Windows][11030]
+- [Input components corresponding to function arguments are now
+  displayed.][11165]
 - [Fixed "rename project" button being broken after not changing project
   name][11103]
 - [Numbers starting with dot (`.5`) are accepted in Numeric Widget][11108]
@@ -41,6 +43,7 @@
 [11001]: https://github.com/enso-org/enso/pull/11001
 [11014]: https://github.com/enso-org/enso/pull/11014
 [11030]: https://github.com/enso-org/enso/pull/11030
+[11165]: https://github.com/enso-org/enso/pull/11165
 [11103]: https://github.com/enso-org/enso/pull/11103
 [11108]: https://github.com/enso-org/enso/pull/11108
 [11056]: https://github.com/enso-org/enso/pull/11056
@@ -60,6 +63,8 @@
 - [Add `Text.to_decimal`.][10874]
 - [Added .floor, .ceil, .trunc to the in-memory `Decimal` column.][10887]
 - [Added vectorized .round to the in-memory `Decimal` column.][10912]
+- [`select_into_database_table` no longer defaults the primary key to the first
+  column][11120]
 
 [10614]: https://github.com/enso-org/enso/pull/10614
 [10660]: https://github.com/enso-org/enso/pull/10660
@@ -71,6 +76,7 @@
 [10874]: https://github.com/enso-org/enso/pull/10874
 [10887]: https://github.com/enso-org/enso/pull/10887
 [10912]: https://github.com/enso-org/enso/pull/10912
+[11120]: https://github.com/enso-org/enso/pull/11120
 
 #### Enso Language & Runtime